Rename source folder to agentMET4FOF
