+++ agentMET4FOF/main_agent_network.py
@@ -1,4 +1,4 @@
-from agentMet4FoF.agents import AgentNetwork
+from agentMET4FOF.agents import AgentNetwork
 
 #Agent modules
 import examples.ZEMA_EMC.zema_agents as zema_agents
