+++ examples/ZEMA_EMC/zema_agents.py
@@ -1,4 +1,4 @@
-from agentMet4FoF.agents import AgentMET4FOF
+from agentMET4FOF.agents import AgentMET4FOF
 from examples.ZEMA_EMC.zema_feature_extract import FFT_BFC, Pearson_FeatureSelection
 
 import numpy as np
