+++ tests/test_memory_monitor_agent.py
@@ -1,6 +1,6 @@
 import pytest
 
-from agentMet4FoF.agents import AgentMET4FOF, AgentNetwork, MonitorAgent
+from agentMET4FOF.agents import AgentMET4FOF, AgentNetwork, MonitorAgent
 import numpy as np
 import pandas as pd
 import time
